+++ .github/workflows/run_docs_check.yml
@@ -33,7 +33,7 @@ jobs:
       uses: swatinem/rust-cache@9d47c6ad4b02e050fd481d890b2ea34778fd09d6
       with:
         save-if: ${{ github.ref == 'refs/heads/main' }}
-    - name: run_tests::run_docs_check::check_docs::lychee_link_check
+    - name: run_docs_checks::lychee_link_check
       uses: lycheeverse/lychee-action@82202e5e9c2f4ef1a55a3d02563e1cb6041e5332
       with:
         args: --no-progress --exclude '^http' './docs/src/**/*'
@@ -44,16 +44,16 @@ jobs:
     - name: steps::install_mold
       run: ./script/install-mold
       shell: bash -euxo pipefail {0}
-    - name: run_tests::run_docs_check::check_docs::install_mdbook
+    - name: run_docs_checks::install_mdbook
       uses: peaceiris/actions-mdbook@ee69d230fe19748b7abf22df32acaa93833fad08
       with:
         mdbook-version: 0.4.37
-    - name: run_tests::run_docs_check::check_docs::build_docs
+    - name: run_docs_checks::build_docs
       run: |
         mkdir -p target/deploy
         mdbook build ./docs --dest-dir=../target/deploy/docs/
       shell: bash -euxo pipefail {0}
-    - name: run_tests::run_docs_check::check_docs::lychee_link_check
+    - name: run_docs_checks::lychee_link_check
       uses: lycheeverse/lychee-action@82202e5e9c2f4ef1a55a3d02563e1cb6041e5332
       with:
         args: --no-progress --exclude '^http' 'target/deploy/docs'
