+++ .github/workflows/run_style_checks.yml
@@ -24,7 +24,7 @@ jobs:
       uses: pnpm/action-setup@fe02b34f77f8bc703788d5817da081398fad5dd2
       with:
         version: '9'
-    - name: run_tests::run_style_checks::check_style::prettier_check_docs
+    - name: run_style_checks::prettier_check_docs
       run: |
         pnpm dlx "prettier@${PRETTIER_VERSION}" . --check || {
           echo "To fix, run from the root of the Zed repo:"
@@ -35,7 +35,7 @@ jobs:
       env:
         PRETTIER_VERSION: 3.5.0
       working-directory: ./docs
-    - name: run_tests::run_style_checks::check_style::prettier_check_default_json
+    - name: run_style_checks::prettier_check_default_json
       run: |
         pnpm dlx "prettier@${PRETTIER_VERSION}" assets/settings/default.json --check || {
           echo "To fix, run from the root of the Zed repo:"
@@ -51,7 +51,7 @@ jobs:
     - name: ./script/check-keymaps
       run: ./script/check-keymaps
       shell: bash -euxo pipefail {0}
-    - name: run_tests::run_style_checks::check_style::check_for_typos
+    - name: run_style_checks::check_for_typos
       uses: crate-ci/typos@80c8a4945eec0f6d464eaf9e65ed98ef085283d1
       with:
         config: ./typos.toml
