+++ tooling/xtask/src/tasks/workflows.rs
@@ -6,7 +6,10 @@ use std::path::Path;
 mod danger;
 mod nix_build;
 mod release_nightly;
+mod run_action_checks;
 mod run_bundling;
+mod run_docs_checks;
+mod run_style_checks;
 
 mod run_tests;
 mod runners;
@@ -25,17 +28,11 @@ pub fn run_workflows(_: GenerateWorkflowArgs) -> Result<()> {
         ("run_bundling.yml", run_bundling::run_bundling()),
         ("release_nightly.yml", release_nightly::release_nightly()),
         ("run_tests.yml", run_tests::run_tests()),
-        (
-            "run_docs_check.yml",
-            run_tests::run_docs_check::run_docs_check(),
-        ),
-        (
-            "run_style_checks.yml",
-            run_tests::run_style_checks::run_style_checks(),
-        ),
+        ("run_docs_check.yml", run_docs_checks::run_docs_checks()),
+        ("run_style_checks.yml", run_style_checks::run_style_checks()),
         (
             "run_action_checks.yml",
-            run_tests::run_action_checks::run_action_checks(),
+            run_action_checks::run_action_checks(),
         ),
         // ("release.yml", release::release()),
     ];
