+++ tooling/xtask/src/tasks/workflows/run_tests.rs
@@ -14,7 +14,7 @@ fn str_vec(values: &'static [&'static str]) -> Vec<String> {
 }
 
 // todo! make this take vec<&str> instead
-fn tests_workflow(paths: &'static [&'static str]) -> Workflow {
+pub(crate) fn tests_workflow(paths: &'static [&'static str]) -> Workflow {
     let paths = str_vec(paths);
     named::workflow()
         // todo! inputs?
@@ -72,157 +72,6 @@ pub(crate) fn run_tests() -> Workflow {
     .add_job(mac_tests.name, mac_tests.job)
     .add_job(migrations.name, migrations.job)
     .add_job(doctests.name, doctests.job)
-}
-
-pub(crate) mod run_action_checks {
-    use super::*;
-
-    pub(crate) fn run_action_checks() -> Workflow {
-        let action_checks = actionlint();
-
-        tests_workflow(&[
-            ".github/workflows/**",
-            ".github/actions/**",
-            ".github/actionlint.yml",
-        ])
-        .add_job(action_checks.name, action_checks.job)
-    }
-
-    fn actionlint() -> NamedJob {
-        const ACTION_LINT_STEP_ID: &'static str = "get_actionlint";
-
-        fn download_actionlint() -> Step<Run> {
-            named::bash("bash <(curl https://raw.githubusercontent.com/rhysd/actionlint/main/scripts/download-actionlint.bash)").id(ACTION_LINT_STEP_ID)
-        }
-
-        fn run_actionlint() -> Step<Run> {
-            named::bash(indoc::indoc! {r#"
-                ${{ steps.get_actionlint.outputs.executable }} -color
-            "#})
-        }
-
-        named::job(
-            release_job(&[])
-                .runs_on(runners::LINUX_SMALL)
-                .add_step(steps::checkout_repo())
-                .add_step(download_actionlint())
-                .add_step(run_actionlint()),
-        )
-    }
-}
-
-pub(crate) mod run_style_checks {
-    use super::*;
-
-    pub(crate) fn run_style_checks() -> Workflow {
-        let style = check_style();
-        tests_workflow(
-            &[], // always check style
-        )
-        .add_job(style.name, style.job)
-    }
-
-    pub(crate) fn check_style() -> NamedJob {
-        fn prettier_check_docs() -> Step<Run> {
-            named::bash(indoc::indoc! {r#"
-                pnpm dlx "prettier@${PRETTIER_VERSION}" . --check || {
-                  echo "To fix, run from the root of the Zed repo:"
-                  echo "  cd docs && pnpm dlx prettier@${PRETTIER_VERSION} . --write && cd .."
-                  false
-                }
-            "#})
-            .working_directory("./docs")
-            .add_env(("PRETTIER_VERSION", "3.5.0"))
-        }
-
-        fn prettier_check_default_json() -> Step<Run> {
-            named::bash(indoc::indoc! {r#"
-                pnpm dlx "prettier@${PRETTIER_VERSION}" assets/settings/default.json --check || {
-                  echo "To fix, run from the root of the Zed repo:"
-                  echo "  pnpm dlx prettier@${PRETTIER_VERSION} assets/settings/default.json --write"
-                  false
-                }
-            "#})
-            .add_env(("PRETTIER_VERSION", "3.5.0"))
-        }
-
-        fn check_for_typos() -> Step<Use> {
-            named::uses(
-                "crate-ci",
-                "typos",
-                "80c8a4945eec0f6d464eaf9e65ed98ef085283d1",
-            ) // v1.38.1
-            .with(("config", "./typos.toml"))
-        }
-
-        named::job(
-            release_job(&[])
-                .runs_on(runners::LINUX_MEDIUM)
-                .add_step(steps::checkout_repo())
-                .add_step(steps::setup_pnpm())
-                .add_step(prettier_check_docs())
-                .add_step(prettier_check_default_json())
-                .add_step(steps::script("./script/check-todos"))
-                .add_step(steps::script("./script/check-keymaps"))
-                .add_step(check_for_typos())
-                // check style steps inlined
-                .add_step(steps::cargo_fmt())
-                .add_step(steps::script("./script/clippy")),
-        )
-    }
-}
-
-pub(crate) mod run_docs_check {
-    use super::*;
-
-    pub(crate) fn run_docs_check() -> Workflow {
-        let docs = check_docs();
-        tests_workflow(&["docs/**"]).add_job(docs.name, docs.job)
-    }
-
-    fn check_docs() -> NamedJob {
-        // todo! would have preferred to just reference the action here while building, but the gh-workflow crate
-        // only supports using repo actions (owner, name, version), not local actions (path)
-        fn lychee_link_check(dir: &str) -> Step<Use> {
-            named::uses(
-                "lycheeverse",
-                "lychee-action",
-                "82202e5e9c2f4ef1a55a3d02563e1cb6041e5332",
-            ) // v2.4.1
-            .add_with(("args", format!("--no-progress --exclude '^http' '{dir}'")))
-            .add_with(("fail", true))
-        }
-
-        fn install_mdbook() -> Step<Use> {
-            named::uses(
-                "peaceiris",
-                "actions-mdbook",
-                "ee69d230fe19748b7abf22df32acaa93833fad08", // v2
-            )
-            .with(("mdbook-version", "0.4.37"))
-        }
-
-        fn build_docs() -> Step<Run> {
-            named::bash(indoc::indoc! {r#"
-                mkdir -p target/deploy
-                mdbook build ./docs --dest-dir=../target/deploy/docs/
-            "#})
-        }
-
-        named::job(
-            release_job(&[])
-                .runs_on(runners::LINUX_LARGE)
-                .add_step(steps::checkout_repo())
-                .add_step(steps::setup_cargo_config(Platform::Linux))
-                // todo(ci): un-inline build_docs/action.yml here
-                .add_step(steps::cache_rust_dependencies())
-                .add_step(lychee_link_check("./docs/src/**/*")) // check markdown links
-                .map(steps::install_linux_dependencies)
-                .add_step(install_mdbook())
-                .add_step(build_docs())
-                .add_step(lychee_link_check("target/deploy/docs")), // check links in generated html
-        )
-    }
 }
 
 pub(crate) fn run_platform_tests(platform: Platform) -> NamedJob {
